consolidate coord directory
