--- pydl/pydlutils/coord/__init__.py
@@ -1,11 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the coord directory of idlutils.
-"""
-from .SDSSMuNu import SDSSMuNu
-from .current_mjd import current_mjd
-from .munu_to_radec import munu_to_radec
-from .radec_to_munu import radec_to_munu
-from .stripe_to_eta import stripe_to_eta
-from .stripe_to_incl import stripe_to_incl
