--- pydl/pydlutils/coord/tests/__init__.py
@@ -1,5 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This is the pydl/pydlutils/coord/tests directory.
-"""
